requirements.txt
# Claude Agent SDK (latest: 0.1.19)
claude-agent-sdk>=0.1.19

# HTTP client for NationBuilder API
httpx>=0.27.0

# Environment variable management
python-dotenv>=1.0.0

# Type hints
typing-extensions>=4.9.0
